requirements.txt
SQLAlchemy==1.4.44
alembic==1.8.1
asyncmy~=0.2.5
python-dateutil~=2.8.2
python-dotenv==0.21.0
timeago~=1.0.16
termcolor~=2.1.1
prettytable==3.5.0
pybeatsaver==0.1.14
pyscoresaber==1.0.9
asyncio~=3.4.3
pyee~=9.0.4
twitchio==2.5.0
discord.py~=2.1.0